change prog1 to progn&find-class to avoid warnings
We get the "can't find type for specializer BAR" warnings because
defmodel puts defclass in a prog1 which is not a top level form.

The advantage of putting defclass in prog1 seems to be it returns the
class object, this can be acheived by using find-class in the tail
position of the progn.

This seems to work on sbcl. Would be interested in your findings